+++ scripts/utils.py
@@ -3,15 +3,11 @@ from pathlib import Path
 import pickle
 from os.path import abspath
 from time import time
-from typing import List, Callable
+from typing import Callable
 
 import numpy as np
 
-from scripts.configs import (
-    DEBUG,
-    DTYPE_FORMAT_F_NUMPY,
-    DTYPE_FORMAT_SIGNED_I_NUMPY,
-)
+from scripts.configs import DEBUG, DTYPE_FORMAT_F_NUMPY, DTYPE_FORMAT_SIGNED_I_NUMPY
 from scripts.utils_numba import is_valid_lat_vec, is_valid_lng_vec
 from timezonefinder import configs
 from timezonefinder.utils_numba import coord2int
@@ -117,35 +113,3 @@ def to_numpy_polygon_repr(coord_pairs, flipped: bool = False) -> np.ndarray:
         y_coords = y_coords[:-1]
     # NOTE: skip expensive validation
     return convert_polygon((x_coords, y_coords), validate=DEBUG)
-
-
-def has_coherent_sequences(lst: List[int]) -> bool:
-    """
-    :return: True if equal entries in the list are not separated by entries of other values
-    """
-    if len(lst) <= 1:
-        return True
-    encountered = set()
-    # at least 2 entries
-    lst_iter = iter(lst)
-    prev = next(lst_iter)
-    for e in lst:
-        if e in encountered:
-            # the entry appeared earlier already
-            return False
-        if e != prev:
-            encountered.add(prev)
-            prev = e
-
-    return True
-
-
-def check_shortcut_sorting(polygon_ids: np.ndarray, all_zone_ids: np.ndarray):
-    # the polygons in the shortcuts are sorted by their zone id (and the size of their polygons)
-    if len(polygon_ids) == 1:
-        # single polygon in the shortcut, no need to check
-        return
-    zone_ids = all_zone_ids[polygon_ids]
-    assert has_coherent_sequences(zone_ids), (
-        f"shortcut polygon ids {polygon_ids} do not have coherent sequences of zone ids: {zone_ids}"
-    )
